add debug of each layer output and fix issues on 64 bit CPU

+++ src/nnom_layers.c
@@ -93,9 +93,9 @@ nnom_layer_t *Conv2D(uint32_t filters, nnom_shape_t k, nnom_shape_t s, nnom_padd
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_conv2d_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
-	comp = (void *)((uint32_t)out + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_conv2d_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
+	comp = (void *)((unsigned long)out + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->super.type = NNOM_CONV_2D;
@@ -159,9 +159,9 @@ nnom_layer_t *Dense(size_t output_unit, const nnom_weight_t *w, const nnom_bias_
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_dense_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
-	comp = (void *)((uint32_t)out + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_dense_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
+	comp = (void *)((unsigned long)out + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->super.type = NNOM_DENSE;
@@ -223,9 +223,9 @@ nnom_layer_t *RNN(nnom_rnn_cell_t *cell, bool return_sequence)
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_rnn_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
-	comp = (void *)((uint32_t)out + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_rnn_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
+	comp = (void *)((unsigned long)out + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->super.type = NNOM_RNN;
@@ -268,8 +268,8 @@ nnom_layer_t *Activation(nnom_activation_t *act)
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_activation_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_activation_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->super.type = NNOM_ACTIVATION;
@@ -335,8 +335,8 @@ nnom_layer_t *Softmax(void)
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->type = NNOM_SOFTMAX;
@@ -365,9 +365,9 @@ nnom_layer_t *MaxPool(nnom_shape_t k, nnom_shape_t s, nnom_padding_t pad_type)
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_maxpool_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
-	comp = (void *)((uint32_t)out + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_maxpool_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
+	comp = (void *)((unsigned long)out + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->super.type = NNOM_MAXPOOL;
@@ -461,8 +461,8 @@ nnom_layer_t *Flatten(void)
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->type = NNOM_FLATTEN;
@@ -490,8 +490,8 @@ nnom_layer_t *Input(nnom_shape_t input_shape, nnom_qformat_t fmt, void *p_buf)
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_io_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_io_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->super.type = NNOM_INPUT;
@@ -542,8 +542,8 @@ nnom_layer_t *Lambda(nnom_status_t (*run)(nnom_layer_t *),
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_lambda_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_lambda_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
 
 	// set buf type.
 	in->type = LAYER_BUF_TEMP;
@@ -585,8 +585,8 @@ static nnom_layer_t *_same_io_shape_base_layer()
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->comp_out_shape = same_io_shape_base_layer_out_shape;
@@ -614,8 +614,8 @@ nnom_layer_t *Concat(int8_t axis)
 		return NULL;
 
 	// distribut the memory to sub handles.
-	in = (void *)((uint32_t)layer + sizeof(nnom_concat_layer_t));
-	out = (void *)((uint32_t)in + sizeof(nnom_layer_io_t));
+	in = (void *)((unsigned long)layer + sizeof(nnom_concat_layer_t));
+	out = (void *)((unsigned long)in + sizeof(nnom_layer_io_t));
 
 	// set type in layer parent
 	layer->super.type = NNOM_CONCAT;
